rename: ChatType을 chat모듈에서 core 모듈로 옮김
