+++ core/src/main/kotlin/kpring/core/chat/chat/dto/request/CreateChatRequest.kt
@@ -2,6 +2,7 @@ package kpring.core.chat.chat.dto.request
 
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
+import kpring.core.chat.model.ChatType
 
 data class CreateChatRequest(
   @field:NotNull
