+++ core/src/main/kotlin/kpring/core/chat/model/ChatType.kt
@@ -1,4 +1,4 @@
-package kpring.core.chat.chat.dto.request
+package kpring.core.chat.model
 
 enum class ChatType(val type: String) {
   Room("Room"),
